+++ changelog.md
@@ -1,6 +1,7 @@
 # Changelog
 
 ### [Latest]
+- Fixed shuffling bug in aux_results and modified treatment of single-track vertices [!259](https://github.com/umami-hep/puma/pull/259)
 - Fixed minor bug in HF vertex merging [!255](https://github.com/umami-hep/puma/pull/255)
 - Fixed removal of reconstructed PVs and modified which vertexing plots are produced given a jet flavour [!253](https://github.com/umami-hep/puma/pull/253)
 - Updated AuxResults class to more closely match Yuma implementation of Results class [!252](https://github.com/umami-hep/puma/pull/252)